+++ src/test/java/chess/domain/position/SquareTest.java
@@ -1,13 +1,9 @@
-package chess.domain.board;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import chess.domain.position.File;
-import chess.domain.position.Rank;
-import chess.domain.position.Square;
 
 public class SquareTest {
     @Test
